requirements.txt
python-dotenv
openai
pyyaml
requests
beautifulsoup4
markdown
arxiv
pymupdf
duckduckgo-search
PyGithub
lightrag
networkx
numpy
pandas
scikit-learn
tqdm
aiohttp
streamlit
nest_asyncio
fastapi
uvicorn[standard]
